Test.Tutorial

Description

Motivation:

So you learned all about parsing, but want to be sure it all works? Welcome to a less beginner-friendly tutorial on testing in Haskell.

Documentation

data SomeParser

Constructors

(EqProp a, Show a, Arbitrary a, CoArbitrary a) => SomeParser (String, Gen (Parse a))

someReturn :: (EqProp a, Show a, Arbitrary a, CoArbitrary a) => a -> SomeParser

rGen :: a -> Gen a

testSomeParser :: (forall a. (EqProp a, Show a, Arbitrary a, CoArbitrary a) => Parse a -> TestBatch) -> SomeParser -> Gen TestBatch

testSomeParser2 :: (forall a b. (EqProp a, Show a, Arbitrary a, CoArbitrary a, EqProp b, Show b, Arbitrary b, CoArbitrary b) => Parse (a, b) -> TestBatch) -> SomeParser -> SomeParser -> Gen TestBatch

testSomeParser3 :: (forall a b c. (EqProp a, Show a, Arbitrary a, CoArbitrary a, EqProp b, Show b, Arbitrary b, CoArbitrary b, EqProp c, Show c, Arbitrary c, CoArbitrary c) => Parse (a, b, c) -> TestBatch) -> SomeParser -> SomeParser -> SomeParser -> Gen TestBatch

prop_parseIsFunctor :: SomeParser -> SomeParser -> SomeParser -> Gen TestBatch

prop_parseIsApplicative :: SomeParser -> SomeParser -> SomeParser -> Gen TestBatch

prop_parseIsMonad :: SomeParser -> SomeParser -> SomeParser -> Gen TestBatch

prop_parseIsMonadFunctor :: SomeParser -> SomeParser -> Gen TestBatch

prop_parseIsAlternative :: SomeParser -> Gen TestBatch

consumes :: Int -> Parse a -> String -> Bool

testCharParser :: (Char -> Bool) -> Parse a -> String -> Bool

Test that a char parser consumes one char if the predicate is true on the first character of the string and 0 otherwise

prop_charP :: (Char -> Bool) -> String -> Bool

prop_char :: Char -> String -> Bool

prop_digit :: String -> Bool

prop_anyChar :: String -> Bool

prop_string :: String -> String -> Bool

prop_notP :: (Arbitrary a, Show a) => Parse a -> String -> Bool

prop_elemP :: String -> String -> Bool

prop_space :: String -> Bool

aParser :: Gen SomeParser

runTests :: IO Bool

Orphan instances

Testable Char

Methods

property :: Char -> Property

Integral a => Show (Expr a)

Methods

showsPrec :: Int -> Expr a -> ShowS

show :: Expr a -> String

showList :: [Expr a] -> ShowS